+++ com.unity.render-pipelines.universal/Runtime/UniversalRenderPipelineCore.cs
@@ -418,14 +418,46 @@ namespace UnityEngine.Rendering.Universal
             return false;
         }
 
-        Comparison<Camera> cameraComparison = (camera1, camera2) => { return (int) camera1.depth - (int) camera2.depth; };
-#if UNITY_2021_1_OR_NEWER
-        void SortCameras(List<Camera> cameras)
+#if ENABLE_VR && ENABLE_VR_MODULE
+        static XR.XRDisplaySubsystem GetXRDisplaySubsystem()
         {
-            if (cameras.Count > 1)
-                cameras.Sort(cameraComparison);
+            XR.XRDisplaySubsystem display = null;
+            SubsystemManager.GetInstances(displaySubsystemList);
+
+            if (displaySubsystemList.Count > 0)
+                display = displaySubsystemList[0];
+
+            return display;
         }
-#else
+
+        // NB: This method is required for a hotfix in Hololens to prevent creating a render texture when using a renderer
+        // with custom render pass.
+        // TODO: Remove this method and usages when we have proper dependency tracking in the pipeline to know
+        // when a render pass requires camera color as input.
+        internal static bool IsRunningHololens(Camera camera)
+        {
+#if PLATFORM_WINRT
+            if (IsStereoEnabled(camera))
+            {
+                var platform = Application.platform;
+                if (platform == RuntimePlatform.WSAPlayerX86 || platform == RuntimePlatform.WSAPlayerARM)
+                {
+                    var displaySubsystem = GetXRDisplaySubsystem();
+                    var subsystemDescriptor = displaySubsystem?.SubsystemDescriptor ?? null;
+                    string id = subsystemDescriptor?.id ?? "";
+
+                    if (id.Contains("Windows Mixed Reality Display"))
+                        return true;
+
+                    if (!XR.WSA.HolographicSettings.IsDisplayOpaque)
+                        return true;
+                }
+            }
+#endif
+            return false;
+        }
+#endif
+
         void SortCameras(Camera[] cameras)
         {
             if (cameras.Length > 1)